refactor: restructure monorepo
