+++ Source/Core/Private/util.cpp
@@ -1,4 +1,4 @@
-#include <glfw3.h>
+#include <GLFW/glfw3.h>
 
 #include "util.h"
 #include "mapping.h"
