+++ Source/Platform/Private/inputManager.cpp
@@ -2,7 +2,7 @@
 #include "stateManager.h"
 #include "util.h"
 
-#include <glfw3.h>
+#include <GLFW/glfw3.h>
 #include <cmath>
 #include <iostream>
