+++ Source/Scripting/Private/solUtil.cpp
